+++ rstar/src/object.rs
@@ -145,7 +145,13 @@ pub trait RTreeObject {
 /// assert!(circle.contains_point(&[1.0, 0.0]));
 /// ```
 pub trait PointDistance: RTreeObject {
-    /// Returns the squared euclidean distance between an object to a point.
+    /// Returns the squared distance between an object and a point.
+    ///
+    /// # Notes
+    /// - While euclidean distance will be the correct choice for most use cases, any distance metric
+    /// fulfilling the [usual axioms](https://en.wikipedia.org/wiki/Metric_space)
+    /// can be used when implementing this method
+    /// - Implementers **must** ensure that the distance metric used matches that of [crate::Envelope::distance_2]
     fn distance_2(
         &self,
         point: &<Self::Envelope as Envelope>::Point,