+++ rstar/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Changed
 - Fixed a stack overflow error in `DrainIterator::next`
+- Clarified that the distance measure in `distance_2` is not restricted to euclidean distance
 
 # 0.11.0