+++ rstar/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Implemented `IntoIter` for `RTree`, i.e. added a owning iterator
 - Added cached envelope bulk load benchmark
 - Implemented `Hash` for `AABB`, `Line`, and `Rectangle`, provided the `Point` used implements `Hash` itself
+- Implemented `Default` for `DefaultParams`
 
 ## Changed
 - Fixed a stack overflow error in `DrainIterator::next`